requirement.txt
Django==4.2
django-ckeditor==6.5.1              
python_dotenv==1.0.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-rest-swagger==2.2.0
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.2
